added a bunch of functionalities, most are autogenerated files

+++ database/migrations/2019_08_21_073631_create_media_table.php
@@ -19,7 +19,7 @@ class CreateMediaTable extends Migration
             $table->string('file_name');
             $table->string('mime_type')->nullable();
             $table->string('disk');
-            $table->unsignedBigInteger('size');
+            $table->unsignedInteger('size');
             $table->json('manipulations');
             $table->json('custom_properties');
             $table->json('responsive_images');
